% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_software_metadata.R
\name{add_codemeta_to_ontosoft}
\alias{add_codemeta_to_ontosoft}
\title{Add codemeta to ontosoft metadata}
\usage{
add_codemeta_to_ontosoft(ontosoft, codemeta)
}
\arguments{
\item{ontosoft}{(ontosoft list) OntoSoft json list object}

\item{codemeta}{(codemeta list) \code{codemeta} json list object}
}
\value{
(ontosoft list) OntoSoft json list object
}
\description{
Add codemeta to ontosoft metadata
}
