% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_software_category.R
\name{align_software_category}
\alias{align_software_category}
\title{Align software category keywords with the IMCR vocabulary}
\usage{
align_software_category(software)
}
\arguments{
\item{software}{(character) Software name}
}
\value{
(list) Updated software JSON as a list object.
  (.json file) Updated software JSON written to the IMCR Portal, and to
  \code{paste0(tempdir(), "/", software, ".json")}.
}
\description{
Align software category keywords with the IMCR vocabulary
}
\examples{
\dontrun{
# Get all IMCR JSON
get_imcr_json()

# Login
login()

# Add software category keywords
update_software_category("add", 'arrow', c('quality control', 'import'))

# Remove software category keywords
update_software_category("remove", 'arrow', c('quality control', 'import'))
}

}
