% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup.R
\name{backup}
\alias{backup}
\title{Back up IMCR JSON to path}
\usage{
backup(path)
}
\arguments{
\item{path}{(character) Path to where IMCR metadata will be written.}
}
\value{
(.json) JSON metadata files written to the specified \code{path}.
}
\description{
Back up IMCR JSON to path
}
