% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_modelo_optimo.R
\name{calcular_modelo_optimo}
\alias{calcular_modelo_optimo}
\title{Remove duplicated strings}
\usage{
calcular_modelo_optimo(datos, nlag)
}
\arguments{
\item{string}{Input vector. Either a character vector, or something
coercible to one.}

\item{...}{Other options used to control matching behavior between duplicate
strings. Passed on to \code{\link[stringi:stri_opts_collator]{stringi::stri_opts_collator()}}.}
}
\value{
A character vector, usually shorter than \code{string}.
}
\description{
La función "calcular_modelo_optimo" tiene como objetivo encontrar
el modelo VECM (Vector Error Correction Model) óptimo para un conjunto
de datos y un número determinado de lags. Para ello, crea una lista
de modelos VECM con diferentes valores de K utilizando la función "ca.jo"
y a continuación crea una lista de modelos VAR (Vector Autoregressive Model)
a partir de los modelos VECM. A continuación, se calcula una matriz de p-values
utilizando la función "serial.test" y se modifican los elementos de la matriz
con una sentencia if-else. Finalmente, se calcula el número óptimo de lags
utilizando la función "which.max" y se calcula el modelo VECM óptimo con la función
"ca.jo" nuevamente. La función retorna una lista con el modelo VECM óptimo y
el número óptimo de lags.
}
\examples{
str_unique(c("a", "b", "c", "b", "a"))

# Use ... to pass additional arguments to stri_unique()
str_unique(c("motley", "mötley", "pinguino", "pingüino"))
str_unique(c("motley", "mötley", "pinguino", "pingüino"), strength = 1)
}
\seealso{
\code{\link[=unique]{unique()}}, \code{\link[stringi:stri_unique]{stringi::stri_unique()}} which this function wraps.
}
