% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanLatex.R
\name{cleanLatex}
\alias{cleanLatex}
\title{Remove LaTeX Auxiliary Files}
\usage{
cleanLatex(folder = getwd(), verbose = FALSE)
}
\arguments{
\item{folder}{The directory to clean. If not specified, it defaults to the current working directory (\code{getwd()}).}

\item{verbose}{Logical; if \code{TRUE}, prints the names of files that are deleted. Default is \code{FALSE}.}
}
\value{
The function returns \code{TRUE} if the files were successfully removed, and \code{FALSE} otherwise.
}
\description{
This function deletes auxiliary files generated after compiling LaTeX in the current working directory.
Common auxiliary file extensions such as \code{.log}, \code{.aux}, \code{.out}, etc., are removed.
}
\details{
The function automatically removes common temporary files generated during the LaTeX compilation process,
including \code{.log}, \code{.aux}, \code{.out}, \code{.bbl}, \code{.toc}, \code{.bak}, and others. It does not prompt the user and will
clean the current directory without confirmation.
}
\examples{
\dontrun{
# Clean the current working directory
cleanLatex()

# Clean a specific directory
cleanLatex("/path/to/latex/files", verbose = TRUE)
}
}
