% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colmat.R
\name{colmat}
\alias{colmat}
\title{Plot color matrix}
\usage{
colmat(x, colframe = FALSE, transpose = TRUE, rowLab = TRUE,
  colLab = TRUE, pal = NULL, lcol = "black", rowLine = -2,
  colLine = -1, las = 2, colSize = 0.7)
}
\arguments{
\item{x}{A matrix with \strong{samples by rows} and 
\strong{variables by columns}. See details.}

\item{colframe}{Logical. Is the dataset already in colors? Default: FALSE}

\item{transpose}{Logical. Transpose matrix? Default: TRUE}

\item{rowLab}{Logical. Label rows? Default: TRUE}

\item{colLab}{Logical. Label columns? Default: TRUE. Printed above plot.}

\item{pal}{Palette used for when \code{colframe = FALSE}. See details.}

\item{lcol}{Line color. Default: 'black'}

\item{rowLine}{\code{line} adjustment for y-axis. See details.}

\item{colLine}{\code{line} adjustment for x-axis. See details.}

\item{las}{\code{las} for plot. Default: 2. See \code{\link[graphics]{par}}.}

\item{colSize}{Size parameter for column labels. See details.}
}
\value{
Plot of a colored table.
}
\description{
\code{colmat} Plot color matrix for factors/categorical data.
}
\details{
Input can either be a matrix where rows are samples and columns
   are variables of factors or colors. For a matrix of factors, define 
   \code{colFrame = FALSE}, \code{makecolors} with palette \code{pal} is
   used to color the matrix. If \code{colFrame = TRUE}, the colors are 
   plotted verbatim.
   
   \code{rowLine} and \code{colLine} are \code{line} arguments for y-lab
   and x-lab respectively. See \code{\link[graphics]{par}}. \code{colSize} 
   is the \code{cex.axis} argument for x-lab.
}
\examples{
# Factor variables
x1 <- matrix(sample(c('Positive','Negative'),50,replace = TRUE),25,2,
             dimnames = list(paste('Sample',1:25,sep='_'),
             c('Mutation_1','Mutation_2')))
x2 <- sample(c('Papillary','Medullary','Mucinous'),25,replace = TRUE)
x <- cbind(x1,Type = x2); rm(x1,x2)
colmat(x)

# Color variables
x <- matrix(sample(c('skyblue3' , 'tan3'), 75, replace = TRUE), 25, 3, 
            dimnames = list(paste('Sample',1:25,sep='_'),
                            c('Var1','Var2','Var3')))
colmat(x)

}
\seealso{
\code{\link{makecolors}}, \code{\link[graphics]{par}}
}
