% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_table.R
\name{combine_table}
\alias{combine_table}
\title{Combine Two Tables with different statistics.}
\usage{
combine_table(tab1, tab2, percent = TRUE)
}
\arguments{
\item{tab1}{A data frame or matrix of the first statistics.}

\item{tab2}{A data frame or matrix of the second statistics to be combined with \code{tab1}.}

\item{percent}{Logical. If \code{TRUE}, formats \code{tab2} as percentages (i.e., appends a \verb{\%} sign). If \code{FALSE}, combines the counts and values without the percentage sign.}
}
\value{
A data frame with the combined counts and percentages (or values).
}
\description{
This function combines two tables: one with raw counts and another with percentages (or other calculated values). It returns a formatted table that merges these values in a "count (percentage)" format.
}
\details{
This function takes two tables, converts them to data frames, and then combines their values in a "count (percentage)" format. If the \code{percent} argument is \code{TRUE}, the values in \code{tab2} are treated as percentages; otherwise, they are combined as raw values.
}
\examples{
tab1<-margin.table(Titanic,c(1,4))
tab2 <- round(prop.table(tab1, margin = 1) * 100, 2)
combine_table(tab1,tab2,)

}
