% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commit_and_tag.R
\name{commit_and_tag}
\alias{commit_and_tag}
\title{Commit Changes and Tag Version in Git}
\usage{
commit_and_tag(version = "1.0.0", message = paste("Release version", version))
}
\arguments{
\item{version}{A string specifying the version number for the tag. Default is "1.0.0".}

\item{message}{A string specifying the commit message. Default is constructed as "Release version " followed by the version number.}
}
\value{
This function returns a message indicating whether the operations were successful or not.
}
\description{
This function commits all changes in the current directory to Git, creates a
tag for the specified version, and pushes both the commit and the tag to the
remote repository.
}
\examples{
\dontrun{
  commit_and_tag("1.0.2")
}

}
