% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_handouts.R
\name{compile_handouts}
\alias{compile_handouts}
\title{Create and Compile Handouts from a List of TeX Files}
\usage{
compile_handouts(file_list, template_path = "handout_tex/handout.tex")
}
\arguments{
\item{file_list}{A list of TeX filenames to be processed and compiled into handouts.}

\item{template_path}{A string specifying the path to the template LaTeX file (default is "handout_tex/handout.tex").}
}
\value{
This function does not return a value but compiles PDFs and deletes the temporary handout TeX files after compilation.
}
\description{
This function processes a list of LaTeX files, creates corresponding handout files using a specified template, compiles them into PDFs, and removes the intermediate handout files after successful compilation.
}
\details{
The function reads a template LaTeX file and loops over a user-provided list of TeX files. For each TeX file, the function generates a corresponding handout file by replacing a placeholder in the template, compiles the file into a PDF using \code{tinytex::latexmk}, and then deletes the handout TeX file once the PDF is generated. If an error occurs during compilation, it skips to the next file.
}
\examples{
\dontrun{
  tex_files <- list.files(pattern = "^Sta.*\\\\.tex")
  compile_handouts(tex_files)
}

}
