% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_texs.R
\name{compile_texs}
\alias{compile_texs}
\title{Compile LaTeX Files}
\usage{
compile_texs(file_list = NULL, clean = TRUE)
}
\arguments{
\item{file_list}{A character vector of LaTeX file names to be compiled. If NULL, all \code{.tex} files in the working directory will be compiled.}

\item{clean}{Logical indicating whether to clean auxiliary files after compilation (default is TRUE).}
}
\value{
A character vector of files that were unsuccessfully compiled.
}
\description{
This function compiles a list of LaTeX (.tex) files using \code{tinytex::latexmk}.
If no file list is provided, it compiles all \code{.tex} files in the current working directory.
Note that files that fail to compile will still be included in the results data frame, marked with \code{success = FALSE}.
}
\examples{
\dontrun{
  failed_files <- compile_texs()
  file_list = list.files(pattern = "^Sta.*tex", ignore.case = TRUE)
  failed_files <- compile_texs(file_list)
}

}
