% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertir_a_csv.R
\name{convertir_a_csv}
\alias{convertir_a_csv}
\title{Imputa datos perdidos en una serie de tiempo}
\usage{
convertir_a_csv(archivos)
}
\arguments{
\item{data}{Una serie de tiempo que contenga datos perdidos}
}
\value{
Un objeto de tipo \code{ts} con los datos imputados.
}
\description{
Esta función recibe una lista de archivos Excel y los convierte a
archivos CSV, omitiendo la primera 5 filas de cada archivo y la última
fila que contenga datos. Para lograr esto, primero se itera sobre la lista
de archivos utilizando un ciclo for. Luego, se usa la función read_excel
del paquete readxl para leer el archivo y almacenarlo en una variable.
A continuación, se utiliza la función nrow para determinar el último índice
que contiene datos en el data frame y se almacena en una variable local.
Finalmente, se usa la función write_csv del paquete readxl para escribir el
archivo CSV, indicando que se deben omitir las primeras 5 filas y la última
fila que contenga datos.
}
\seealso{
\code{\link[=unique]{unique()}}, \code{\link[stringi:stri_unique]{stringi::stri_unique()}} which this function wraps.
}
