% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpFun.R
\name{count_percent}
\alias{count_percent}
\title{Calculate Counts and Percentages of Factor Levels}
\usage{
count_percent(x, digits = 2)
}
\arguments{
\item{x}{A factor or a categorical variable.}

\item{digits}{An integer indicating the number of decimal places to which the
percentages will be rounded. Defaults to 2.}
}
\value{
A character vector where each element corresponds to a level of \code{x}.
Each element is formatted as 'count (percentage\%)', where 'count' is the number of occurrences of the level
and 'percentage' is the percentage of the total, rounded to \code{digits} decimal places.
}
\description{
This function computes the counts and percentages for each level of a factor or categorical variable.
It returns a character vector with each level's count and its percentage of the total, formatted to a specified number of decimal places.
}
\examples{
levels= LETTERS[1:3]
factor1 <- levels\%>\%sample(size = 10, replace = TRUE)\%>\%factor()
count_percent(factor1,1)

}
