% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_txt.R
\name{cut_txt}
\alias{cut_txt}
\title{Cut text file into chunks with specified number of chapters}
\usage{
cut_txt(file, marker = "<e7><ac><ac>.*<e7><ab><a0>", num_chapters = 25)
}
\arguments{
\item{file}{The path to the text file to be cut.}

\item{marker}{The regular expression pattern to identify chapters in the text file.}

\item{num_chapters}{The number of chapters to include in each chunk. Default is 25.}
}
\value{
NULL (The function writes output files.)
}
\description{
This function reads a text file and splits it into chunks where each chunk contains a specified number of chapters.
Chapters are identified by markers like "\if{html}{\out{<e7>}}\if{html}{\out{<ac>}}\if{html}{\out{<ac>}}*\if{html}{\out{<e7>}}\if{html}{\out{<ab>}}\if{html}{\out{<a0>}}".
}
\examples{
\dontrun{
cut_txt(file = "your_file.txt", marker = "<e7><ac><ac>.*<e7><ab><a0>", num_chapters = 25)
}

}
