% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_software.R
\name{delete_software}
\alias{delete_software}
\title{Delete software from the IMCR Portal}
\usage{
delete_software(name)
}
\arguments{
\item{name}{(character) Name of software}
}
\value{
(message) Status message of DELETE operation.
}
\description{
Delete software from the IMCR Portal
}
\examples{
\dontrun{
get_imcr_json()
login()
delete_software("arrow")
}

}
