% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpsd.R
\name{dpsd}
\alias{dpsd}
\title{Differential probes by standard deviation}
\usage{
dpsd(x, n, .order = TRUE)
}
\arguments{
\item{x}{Matrix of numbers with columns indicating samples and rows
indicating features.}

\item{n}{Number of features to choose.}

\item{.order}{Logical. Return features in decreasing SD values.}
}
\value{
A matrix of nrow = n features with highest sd.
}
\description{
\code{dpsd} Find most variable features by standard deviation.
}
\details{
Identifies the most variable features across samples by
   standard deviation (sd). The \code{n} features with highest sd
   is returned as a matrix.
}
\examples{
x1 <- matrix(rnorm(1000), nrow = 50)
rownames(x1) = paste('LowSD', 1:50, sep='_')
x2 <- matrix(rnorm(1000, sd = 3), nrow = 50)
rownames(x2) = paste('HighSD', 1:50, sep='_')
dpsd(rbind(x1,x2),20)

}
