% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_normal.R
\name{draw_normal}
\alias{draw_normal}
\title{Draw a normal distribution plot using ggplot2.}
\usage{
draw_normal(mu = 0, sd = 1, color_density = "blue")
}
\arguments{
\item{mu}{Mean of the normal distribution.}

\item{sd}{Standard deviation of the normal distribution.}
}
\value{
A ggplot2 object representing the normal distribution plot.
}
\description{
This function creates a ggplot2 plot representing the probability density function (PDF)
of a normal distribution with specified mean and standard deviation.
}
\examples{
draw_normal(mu = 0, sd = 1)

}
