% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_links.R
\name{fun_links}
\alias{fun_links}
\title{Generate Links to Function Documentation for packagedown}
\usage{
fun_links(function_names, package_name = "toolkit")
}
\arguments{
\item{function_names}{A character vector of function names.}

\item{package_name}{A string specifying the name of the package (default is "toolkit").}
}
\value{
A single concatenated string containing markdown links to the function documentation.
}
\description{
This function takes a list of function names and generates a markdown
formatted string with links to their documentation.
}
\examples{
function_list <- c("compile_rnws", "cleanFolder", "knit_rnws")
result <- fun_links(function_list, package_name = "toolkit")
cat(result)
}
