% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_software.R
\name{get_all_software}
\alias{get_all_software}
\title{Get all software metadata of an OntoSoft portal}
\usage{
get_all_software(url)
}
\arguments{
\item{url}{(character) OntoSoft portal URL.}
}
\value{
(list) Named list of software JSON metadata.
}
\description{
Get all software metadata of an OntoSoft portal
}
\examples{
\dontrun{
json <- get_all_software('http://imcr.ontosoft.org/repository/software')
}

}
