% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_broader_terms.R
\name{get_broad_terms}
\alias{get_broad_terms}
\title{Get broad terms of an IMCR Vocabulary keyword}
\usage{
get_broad_terms(term)
}
\arguments{
\item{term}{(character) IMCR Vocabulary keyword}
}
\value{
(list) Named list of terms (including searched term) and corresponding
  IDs. NA is returned for non-existant terms.
}
\description{
Get broad terms of an IMCR Vocabulary keyword
}
\examples{
\dontrun{
get_broad_terms("workflow planning")
}

}
