% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_broader_terms.R
\name{get_broader_terms}
\alias{get_broader_terms}
\title{Get broader terms of an IMCR Vocabulary keyword}
\usage{
get_broader_terms(term)
}
\arguments{
\item{term}{(character) IMCR Vocabulary keyword}
}
\value{
(list) Named list of terms (including searched term) and corresponding
  IDs. NA is returned for non-existant terms.
}
\description{
Get broader terms of an IMCR Vocabulary keyword
}
\examples{
\dontrun{
get_broader_terms("workflow planning")
}

}
