% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ids.R
\name{get_ids}
\alias{get_ids}
\title{Get all software IDs of an OntoSoft Portal.}
\usage{
get_ids(url)
}
\arguments{
\item{url}{(character) OntoSoft portal URL}
}
\value{
(list) Named list of software IDs. Names are softare titles.
}
\description{
Get all software IDs of an OntoSoft Portal.
}
\examples{
\dontrun{
ids <- get_ids('http://imcr.ontosoft.org/repository/software')
}

}
