% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_json.R
\name{get_json}
\alias{get_json}
\title{Get JSON metadata for all portal software}
\usage{
get_json(ids)
}
\arguments{
\item{ids}{(character) OntoSoft software identifiers.}
}
\value{
(list) JSON metadata for each identifier.
}
\description{
Get JSON metadata for all portal software
}
\examples{
\dontrun{
ids <- get_ids('http://imcr.ontosoft.org/repository/software')
json <- get_json(ids)
}

}
