% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_software_category.R
\name{get_software_category}
\alias{get_software_category}
\title{Get Software Category}
\usage{
get_software_category(json)
}
\arguments{
\item{json}{(list) Software metadata in JSON format}
}
\value{
(list) Software keywords
}
\description{
Get software category keywords.
}
\examples{
\dontrun{
json <- get_json('http://imcr.ontosoft.org/repository/software')
parse_keywords(json)
}

}
