% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_SoftwareCategory.R
\name{get_SoftwareCategory}
\alias{get_SoftwareCategory}
\title{Get Software Category}
\usage{
get_SoftwareCategory(json)
}
\arguments{
\item{json}{(list) Software metadata in JSON format}
}
\value{
(list) Software keywords
}
\description{
Get Software Category
}
\examples{
\dontrun{
json <- get_json('http://imcr.ontosoft.org/repository/software')
parse_keywords(json)
}

}
