% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_software_metadata.R
\name{gh_to_codemeta}
\alias{gh_to_codemeta}
\title{Get \code{codemeta} from GitHub URL}
\usage{
gh_to_codemeta(github.url)
}
\arguments{
\item{github.url}{(character) GitHub URL}
}
\value{
(codemeta list) A \code{codemeta} list object
}
\description{
Get \code{codemeta} from GitHub URL
}
