% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_commit_push.R
\name{git_commit_push}
\alias{git_commit_push}
\title{Commit Changes and Push to Git}
\usage{
git_commit_push(message = "")
}
\arguments{
\item{message}{A string specifying the commit message.}
}
\value{
This function returns a message indicating whether the operations were successful or not.
}
\description{
This function commits all changes in the current directory to Git and pushes the commit to the remote repository.
}
\examples{
\dontrun{
  git_commit_push("")
}

}
