% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_tag_push.R
\name{git_tag}
\alias{git_tag}
\title{Tag Version in Git}
\usage{
git_tag(version = "1.0.0", message = paste("Release version", version))
}
\arguments{
\item{version}{A string specifying the version number for the tag. Default is "1.0.0".}

\item{message}{A string specifying the tag message. Default is constructed as "Release version " followed by the version number.}
}
\value{
This function returns a message indicating whether the tag operation was successful or not.
}
\description{
This function creates a tag for the specified version and pushes the tag to the remote repository.
}
\examples{
\dontrun{
  git_tag("1.0.2")
}

}
