% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_images.R
\name{insert_images}
\alias{insert_images}
\title{Insert multiple images into R Markdown documents}
\usage{
insert_images(image_names, default_folder = NULL)
}
\arguments{
\item{image_names}{Character vector of image filenames or paths. Files
without directory paths will be relative to the default folder.}

\item{default_folder}{Optional folder path to override the global default.
If NULL (default), uses the folder set by \code{set_default_image_folder()}.}
}
\value{
Output from \code{knitr::include_graphics()} containing the existing
image files.
}
\description{
Insert multiple images into R Markdown documents
}
\details{
This function is designed to be used within R Markdown chunks.
It applies the default folder to image names that don't contain path
separators (/ or \\), checks for file existence, and includes only
existing files in the output. Missing files generate warnings but don't
stop execution.
}
\examples{
\dontrun{
# After setting default folder
set_default_image_folder("images/")
insert_images(c("plot1.png", "plot2.png"))

# With mixed paths
insert_images(c("plot1.png", "../other/plot2.png"))

# Override default folder
insert_images(c("plot1.png"), default_folder = "alternative/path")
}

}
\seealso{
\code{\link{set_default_image_folder}}
}
