% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_software_metadata.R
\name{join_rpkg_github}
\alias{join_rpkg_github}
\title{Join \code{codemeta} of R Package Description and GitHub}
\usage{
join_rpkg_github(cm.rpkg, cm.github)
}
\arguments{
\item{codemeta.rpkg}{(codemeta list) \code{codemeta} of R Package Description}

\item{codemeta.github}{(codemeta list) \code{codemeta} of GitHub}
}
\value{
(codemta list) \code{codemeta} list object
}
\description{
Join \code{codemeta} of R Package Description and GitHub
}
