% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_authors.R
\name{list_creators}
\alias{list_creators}
\title{List software creators}
\usage{
list_creators(json)
}
\arguments{
\item{json}{(list) Software metadata in JSON format}
}
\value{
(list) Software creators
}
\description{
List software creators
}
\examples{
\dontrun{
json <- get_json('http://imcr.ontosoft.org/repository/software')
list_creators(json)
}

}
