% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_keywords.R
\name{list_keywords}
\alias{list_keywords}
\title{List kewords}
\usage{
list_keywords(json)
}
\arguments{
\item{json}{(list) Software metadata in JSON format}
}
\value{
(list) Software keywords
}
\description{
List kewords
}
\examples{
\dontrun{
json <- get_json('http://imcr.ontosoft.org/repository/software')
parse_keywords(json)
}

}
