% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_languages.R
\name{list_language}
\alias{list_language}
\title{List implementation languages}
\usage{
list_language(json)
}
\arguments{
\item{json}{(list) Software metadata in JSON format}
}
\value{
(list) Software languages
}
\description{
List implementation languages
}
\examples{
\dontrun{
json <- get_json('http://imcr.ontosoft.org/repository/software')
list_languages(json)
}

}
