% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{login}
\alias{login}
\title{Login}
\usage{
login(user.name = NULL, user.pass = NULL)
}
\arguments{
\item{user.name}{(character) OntoSoft user name. Default (NULL) queries for this
information in the console.}

\item{user.pass}{(character) OntoSoft user password. Default (NULL) queries for this
information in the console.}
}
\value{
(character) Session string.
}
\description{
Login
}
\examples{
\dontrun{
# Enter credentials to function
login('myname', 'mypass')
# Have credentials requested in the console
login()
}

}
