% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makecolors.R
\name{makecolors}
\alias{makecolors}
\title{Make color labels for factors/categorical data.}
\usage{
makecolors(object, pal = cbbPalette, keep.order = TRUE)
}
\arguments{
\item{object}{A matrix of vector that can be coerced into factors.}

\item{pal}{A vector of colors.}

\item{keep.order}{Logical. Keep original order of non-factor strings.}
}
\value{
A list containing:
   \describe{
     \item{Legend:}{A unique vector/color pair for easy legend plotting.}
     \item{Colors:}{A matrix or vector of colors for plotting.}
   }
}
\description{
\code{makecolors} Generates color labels and legends for factors.
}
\details{
The function accepts either a vector that can be coerced into 
factors or a matrix where each \strong{column} is used to generate color 
labels.

Can also be generate unique \code{pch} labels (see \bold{Examples}).
}
\examples{
x <- factor(sample(c('a','b','c'), 30, replace = TRUE))
colset <- makecolors(x)
xnum <- jitter(as.numeric(x))
plot(xnum, col = colset$Colors, pch = 16)
legend('topright', legend = names(colset$Legend), col = colset$Legend, pch = 16)

pchset <- makecolors(x, pal = c(16,17,18))$Colors
plot(xnum, col = colset$Colors, pch = pchset)
}
