% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpFun.R
\name{meanSD}
\alias{meanSD}
\title{Calculate Mean and Standard Deviation, and Format the Output}
\usage{
meanSD(x, digits = 2)
}
\arguments{
\item{x}{A numeric vector for which the mean and standard deviation are to be computed.}

\item{digits}{An integer indicating the number of decimal places to which the
results will be rounded. Defaults to 2.}
}
\value{
A character string containing the mean and standard deviation of \code{x},
formatted as 'mean (standard deviation)', each rounded to \code{digits} decimal places.
}
\description{
This function computes the mean and standard deviation of a numeric vector
and returns a formatted string with the mean and standard deviation, rounded to
a specified number of decimal places.
}
\examples{
meanSD(c(10.5, 5.3, 7.8), digits = 3)
}
