% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midPoints.R
\name{midPoints}
\alias{midPoints}
\title{Calculate cumulative midpoints}
\usage{
midPoints(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A numeric vector of length \code{length(x) - 1}
}
\description{
\code{midPoints} Calculate cumulative midpoints of a numeric vector.
}
\details{
Calculates the cumulative midpoints of a numeric vector. To
   be used in plotting functions, especially for Manhattan plots.
}
\examples{
midPoints(rep(1, 10))

}
