% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthday.R
\name{monthday}
\alias{monthday}
\title{Extract and Format Month and Day}
\usage{
monthday(date)
}
\arguments{
\item{date}{A Date object or a character string convertible to a Date. The date to extract month and day from.}
}
\value{
A character string combining the abbreviated month and day (e.g., "Jan 1").
}
\description{
This function takes a date input, extracts the month and day, and formats them as a string.
}
\examples{
monthday(as.Date("2024-01-01"))
monthday(Sys.Date())
}
