% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myheatmap2.R
\name{myheatmap2}
\alias{myheatmap2}
\title{Personalized heatmap.2}
\usage{
myheatmap2(x, col = colorRampPalette(c("blue", "black", "yellow"))(50), ...)
}
\arguments{
\item{x}{Numeric matrix of values to be plotted. Data.frame will be coerced.}

\item{col}{Defaults to a blue-black-yellow color scheme with 50 shades}

\item{...}{Other arguments to be passed on to \code{heatmap.2}}
}
\value{
An object of the class list from \code{heatmap.2}.
}
\description{
\code{myheatmap2} Runs \code{heatmap.2} with personalized color scheme.
}
\details{
Turns of \code{density.info} and \code{trace} and sets the
color scheme to blue-black-yellow. Other arguments can be passed onto
\code{heatmap.2}. Most commonly used are \code{ColSideCol} and 
\code{RowSideCol}.
}
\seealso{
\code{\link[gplots]{heatmap.2}}
}
