% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_languages.R
\name{plot_languages}
\alias{plot_languages}
\title{Plot software languages}
\usage{
plot_languages(languages)
}
\arguments{
\item{languages}{(list) Software languages}
}
\value{
(.png) Pie chart of input implementation languages
}
\description{
Plot software languages
}
\examples{
\dontrun{
json <- get_json('http://imcr.ontosoft.org/repository/software')
list_creators(json)
}

}
