% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmatrix.R
\name{pmatrix}
\alias{pmatrix}
\title{Print a Matrix in LaTeX pmatrix Format}
\usage{
pmatrix(matr, digits = 2)
}
\arguments{
\item{matr}{A numeric matrix to be printed.}

\item{digits}{An integer specifying the number of decimal places to round the elements of the matrix. Default is 2.}
}
\value{
The function does not return a value; it prints the matrix in LaTeX \code{pmatrix} format.
}
\description{
This function prints a given matrix in LaTeX \code{pmatrix} format, rounded to a specified number of decimal places.
}
\examples{
mat <- matrix(c(1.123456, 2.654321, 3.987654), nrow = 3)
pmatrix(mat, digits = 3)

}
