% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointbox.R
\name{pointbox}
\alias{pointbox}
\title{Boxplots with points}
\usage{
pointbox(value, grp, b_col = NULL, s_col = NULL, s_pch = 16,
  s_jitter = NULL, showN = TRUE, heightN = NULL, stat_test = TRUE,
  y_range = NULL, s_outline = TRUE, ...)
}
\arguments{
\item{value}{Numeric variable for N measurements.}

\item{grp}{Classification of measurements.}

\item{b_col}{Boxplot color. Default: NULL (black)}

\item{s_col}{Scatterplot color. Default: NULL (black). If specified, length of N, 
as each point can be colored separately.}

\item{s_pch}{Point type. Default: 16, filled circle.}

\item{s_jitter}{Jitter factor for points.}

\item{showN}{Logical. Show number of samples in each classification.}

\item{heightN}{Manually set height for N if showN = TRUE.}

\item{stat_test}{Logical. Perform statistical test? Currently, only one-sided
anova is implemented.}

\item{y_range}{Manually specify y-axis range.}

\item{s_outline}{Logical. Outline points.}

\item{...}{Arguments passed on to par. See \code{\link[graphics]{par}}.}
}
\value{
Plot of a colored table.
}
\description{
\code{pointbox} Plot boxplots with points (scatterplot).
\code{pointbox_label} Plot boxplots with points and labels each point.
}
\details{
Input can either be a matrix where rows are samples and columns
   are variables of factors or colors. For a matrix of factors, define 
   \code{colFrame = FALSE}, \code{makecolors} with palette \code{pal} is
   used to color the matrix. If \code{colFrame = TRUE}, the colors are 
   plotted verbatim.
   
   \code{rowLine} and \code{colLine} are \code{line} arguments for y-lab
   and x-lab respectively. See \code{\link[graphics]{par}}. \code{colSize} 
   is the \code{cex.axis} argument for x-lab.
}
\examples{
# Factor variables
x1 <- matrix(sample(c('Positive','Negative'),50,replace = TRUE),25,2,
             dimnames = list(paste('Sample',1:25,sep='_'),
             c('Mutation_1','Mutation_2')))
x2 <- sample(c('Papillary','Medullary','Mucinous'),25,replace = TRUE)
x <- cbind(x1,Type = x2); rm(x1,x2)
colmat(x)

# Color variables
x <- matrix(sample(c('skyblue3' , 'tan3'), 75, replace = TRUE), 25, 3, 
            dimnames = list(paste('Sample',1:25,sep='_'),
                            c('Var1','Var2','Var3')))
colmat(x)

}
\seealso{
\code{\link[graphics]{par}}
}

