% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointbox_label.R
\name{pointbox_label}
\alias{pointbox_label}
\title{Boxplots with points and labels}
\usage{
pointbox_label(value, grp, label_names = NULL, b_col = NULL, s_col = NULL,
  s_pch = 16, s_jitter = NULL, showN = TRUE, heightN = NULL,
  stat_test = TRUE, y_range = NULL, ...)
}
\arguments{
\item{value}{Numeric variable for N measurements.}

\item{grp}{Classification of measurements.}

\item{label_names}{Character vector of length value. Use NA to hide specific names.}

\item{b_col}{Boxplot color. Default: NULL (black)}

\item{s_col}{Scatterplot color. Default: NULL (black). If specified, length of N, 
as each point can be colored separately.}

\item{s_pch}{Point type. Default: 16, filled circle.}

\item{s_jitter}{Jitter factor for points.}

\item{showN}{Logical. Show number of samples in each classification.}

\item{heightN}{Manually set height for N if showN = TRUE.}

\item{stat_test}{Logical. Perform statistical test? Currently, only one-sided
anova is implemented.}

\item{y_range}{Manually specify y-axis range.}

\item{...}{Arguments passed on to par. See \code{\link[graphics]{par}}.}

\item{s_outline}{Logical. Outline points.}
}
\value{
Boxplot with horizontally jittered points.
}
\description{
\code{pointbox_label} Plot boxplots with points and labels each point.
}
\details{
Colors for boxes and points are defined separately to allow for
   fine-tuned user control.
}
\examples{
# Make data
d <- c(rnorm(n = 10, mean = 1, sd = 0.5),
       rnorm(n = 10, mean = 5, sd = 0.5))
l <- paste0('L', 1:20)
g <- c(rep('a', 10), rep('b', 10))
pointbox_label(value = d, grp = g, label_names = l)

}
\seealso{
\code{\link[graphics]{par}}
}
