% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_kable.R
\name{print_kable}
\alias{print_kable}
\title{Print and Format a Styled Table for HTML or LaTeX Output}
\usage{
print_kable(
  x,
  num_col = 1,
  rowname_label = NA,
  caption = NA,
  digits = 2,
  fontsize = 16,
  big.mark = ",",
  na_str = "",
  booktabs = TRUE,
  escape = TRUE,
  format.args = list(decimal.mark = ".", big.mark = ","),
  format,
  ...
)
}
\arguments{
\item{x}{The  matrix or data frame to be formatted as a table.}

\item{num_col}{An integer specifying the number of columns to split the output table into (default: 1).}

\item{rowname_label}{A character string specifying the column name for the row names; if not provided, the row names will not be printed (default: NA).}

\item{caption}{A character string specifying the caption for the table (default: '').}

\item{digits}{An integer indicating the number of digits to display (default: 2).}

\item{fontsize}{Font size for the table (default: 9).}

\item{big.mark}{A character string used as the thousands separator for numbers (default: ',').}

\item{na_str}{A string used to replace missing values (default: '').}

\item{booktabs}{Logical indicating whether to use the booktabs formatting (default: TRUE).}

\item{escape}{Logical indicating whether to escape special characters in the table (default: TRUE).}

\item{format.args}{A list of additional formatting arguments to customize number representation.}

\item{format}{A character string specifying the output format: either "latex" or "html".}

\item{...}{Additional arguments to be passed to the kbl function.}
}
\value{
A styled and formatted table in the specified format (LaTeX or HTML).
}
\description{
This function prints and formats a styled table in either HTML or LaTeX format using kableExtra.
}
\examples{
dt = mtcars[1:10, 1:3]
print_flextable(dt, num_col = 2, rowname_label = 'car',format='html')

}
