% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_table.R
\name{print_table}
\alias{print_table}
\title{Print a Styled Table for HTML or LaTeX Output}
\usage{
print_table(
  df,
  caption = "",
  digits = 2,
  print_rownames = FALSE,
  fontsize = 9,
  big.mark = ",",
  na_str = "",
  ...
)
}
\arguments{
\item{df}{The data frame to be formatted as a table.}

\item{caption}{A character string specifying the caption for the table (default: '').}

\item{digits}{An integer indicating the number of digits to display (default: 2).}

\item{print_rownames}{Logical value indicating whether to print row names (default: FALSE).}

\item{fontsize}{Font size for the table (default: 9).}

\item{big.mark}{A character string used as the big mark separator for numbers (default: ',').}

\item{na_str}{A string used to replace missing values (default: '').}

\item{...}{Additional arguments to be passed to the formatting functions.}
}
\value{
A styled table in HTML or LaTeX format, depending on the document type.
}
\description{
This function prints a styled table in either HTML or LaTeX format, depending on the output document type.
}
\examples{
head(mtcars) \%>\% print_table()

}
