% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put_software.R
\name{put_software}
\alias{put_software}
\title{Update software in the IMCR Portal}
\usage{
put_software(name, path)
}
\arguments{
\item{name}{(character) File name of software JSON (without file extension).}

\item{path}{(character) Path to software JSON.}
}
\value{
(message) Status message of PUT operation.
}
\description{
Update software in the IMCR Portal
}
\examples{
\dontrun{
login()
put_software("arrow", "/path/to/software/json")
}

}
