% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpFun.R
\name{range}
\alias{range}
\title{Get range of a vector}
\usage{
range(x, digits = 2)
}
\arguments{
\item{digits}{number of digits after the decimal point. Default is 2.}
}
\description{
Get range of a vector
}
\examples{
range(rnorm(100),3)
}
