% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dict.R
\name{read_dict}
\alias{read_dict}
\title{Read dictionary}
\usage{
read_dict(dict, genome = "hs37d5")
}
\arguments{
\item{dict}{.dict file containing reference information to read.}

\item{genome}{Genome name. ['hs37d5']}
}
\value{
A seqinfo object to be used with GenomicRanges.
}
\description{
\code{read_dict} Read .dict files generated using Picard tools as seqinfo objects.
}
\details{
Reads an input .dict file as a seqinfo object to be used in
GenomicRanges.
}
\examples{
x1 <- matrix(rnorm(1000), nrow = 50)
rownames(x1) = paste('LowSD', 1:50, sep='_')
x2 <- matrix(rnorm(1000, sd = 3), nrow = 50)
rownames(x2) = paste('HighSD', 1:50, sep='_')
dpsd(rbind(x1,x2),20)

}
