% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_punctuation.R
\name{replace_punctuation}
\alias{replace_punctuation}
\title{Replace Chinese Punctuation with English Equivalents}
\usage{
replace_punctuation(filename)
}
\arguments{
\item{filename}{A character string representing the file name to be processed.}
}
\value{
A character string with the modified file name.
}
\description{
This function takes a file name as input and replaces specific Chinese punctuation marks
with their English equivalents. It also converts Traditional Chinese characters to Simplified
Chinese characters.
}
\examples{
updated_name <- replace_punctuation("A，B.m4a")

}
