% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_hook_set.R
\name{rmd_hook_set}
\alias{rmd_hook_set}
\title{Set custom knit hooks for R Markdown}
\usage{
rmd_hook_set()
}
\value{
None
}
\description{
This function sets custom knit hooks for R Markdown documents to customize behavior during knitting.
}
\details{
The function sets inline knit hook to format numeric values and par hook to customize plotting parameters.
}
\examples{
rmd_hook_set()

}
