% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_knit_hook_set.R, R/rmd_knit_hook_set<.R
\name{rmd_knit_hook_set}
\alias{rmd_knit_hook_set}
\title{Set custom knit hooks for R Markdown}
\usage{
rmd_knit_hook_set()

rmd_knit_hook_set()
}
\value{
None

None
}
\description{
This function sets custom knit hooks for R Markdown documents to customize behavior during knitting.

This function sets custom knit hooks for R Markdown documents to customize behavior during knitting.
}
\details{
The function sets inline knit hook to format numeric values and par hook to customize plotting parameters.

The function sets inline knit hook to format numeric values and par hook to customize plotting parameters.
}
\examples{
rmd_knit_hook_set()

rmd_knit_hook_set()

}
