% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_opts_set.R
\name{rmd_opts_set}
\alias{rmd_opts_set}
\title{Set default options for code chunks in R Markdown}
\usage{
rmd_opts_set()
}
\value{
None
}
\description{
This function sets default options for code chunks in R Markdown documents based on the output format.
}
\details{
The function determines the output format (LaTeX or HTML) and sets appropriate chunk options.
For LaTeX output, it sets options such as comment style, output width, figure settings, caching,
and code styling. For HTML output, it sets similar options tailored for HTML rendering.
}
\examples{
rmd_opts_set()

}
