% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnw.verb.hook.R
\name{rnw.verb.hook}
\alias{rnw.verb.hook}
\title{Format Code Output with Custom Font Size in LaTeX}
\usage{
rnw.verb.hook(x, options)
}
\arguments{
\item{x}{A character string containing the code to be formatted.}

\item{options}{A list of chunk options, including \code{outputsize}, which sets the font size
in the Verbatim environment.}
}
\value{
A character vector containing the LaTeX-formatted Verbatim environment with the specified font size.
}
\description{
This function wraps code in a LaTeX \code{Verbatim} environment for display in LaTeX documents.
The font size of the code output can be customized by setting the \code{outputsize} parameter
in the chunk options.
}
\details{
\itemize{
\item You can set the font size for the Verbatim environment by specifying \code{outputsize} in the chunk options.
For example, \code{outputsize = 'small'} will use \verb{\\small} for the font size.
\item Valid font size values in LaTeX include options like \code{tiny}, \code{scriptsize}, \code{footnotesize}, \code{small},
\code{normalsize}, \code{large}, \code{Large}, \code{LARGE}, \code{huge}, etc.
}
}
\examples{
# Setting output font size to small in a chunk:
# rnw.verb.hook("my code", list(outputsize = "small"))
}
\keyword{internal}
