% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnw_hook_set.R
\name{rnw_hook_set}
\alias{rnw_hook_set}
\title{Set Custom Hooks for knitr in RNW Files}
\usage{
rnw_hook_set()
}
\description{
The \code{rnw_hook_set} function customizes knitr hooks to control the formatting of code chunks,
output, inline values, and plot parameters in \code{.rnw} files. It is designed to improve
the appearance of compiled LaTeX documents by setting specific knitr hooks.
}
\details{
This function configures several hooks:
\itemize{
\item \code{chunk}: Defines the formatting of chunks via the \code{rnw.chunk.hook} function.
\item \code{output}: Sets output formatting via the \code{rnw.output.hook} function.
\item \code{inline}: Formats inline numerical output to three decimal places.
\item \code{crop}: Uses \code{hook_pdfcrop} to crop figures if \code{crop} is set to \code{TRUE} in chunk options.
\item \code{par}: Controls graphical parameters like margins and text sizes for plots, with a light gray background.
}
}
\examples{
\dontrun{
  rnw_hook_set() # Set hooks for knitr
}
}
