% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_flextable.R
\name{save_flextable}
\alias{save_flextable}
\title{Save a Flextable as a PDF}
\usage{
save_flextable(ft, path, width = 7, height = NULL)
}
\arguments{
\item{ft}{A flextable object to be saved as a PDF.}

\item{path}{A character string specifying the file path for the output PDF.}

\item{width}{A numeric value specifying the width of the PDF in inches (default: 7).}

\item{height}{A numeric value specifying the height of the PDF in inches. If NULL, the height is calculated
based on the aspect ratio of the flextable.}
}
\value{
NULL. The function saves the flextable as a PDF file at the specified path.
}
\description{
This function saves a flextable object as a PDF file. The height of the table
is calculated based on its aspect ratio if not explicitly provided.
}
\examples{
ft <- print_flextable(mtcars[1:10, 1:3], num_col = 2, rowname_label = 'car')
\dontrun{
 save_flextable(ft, "table.pdf", width = 6, height = 4)
}

}
