% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_default_image_folder.R
\name{set_default_image_folder}
\alias{set_default_image_folder}
\title{Set a global default folder for image files}
\usage{
set_default_image_folder(folder_path)
}
\arguments{
\item{folder_path}{Path to the default image folder. If the folder does not
exist, a warning will be issued but the setting will still be applied.}
}
\value{
No return value, called for side effects. Sets the global option
\code{default_image_folder} and displays a confirmation message.
}
\description{
Set a global default folder for image files
}
\details{
This function sets a global default folder that will be used by
\code{insert_figures()} when no specific folder is provided. The function
checks if the folder exists and issues a warning if it doesn't, but still
sets the option to allow for folder creation later.
}
\examples{
\dontrun{
set_default_image_folder("images/")
}

}
