% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_ggplot_theme.R
\name{set_ggplot_theme}
\alias{set_ggplot_theme}
\title{Set custom theme for ggplot2 plots}
\usage{
set_ggplot_theme()
}
\value{
None
}
\description{
This function sets a custom theme for ggplot2 plots with a light background color.
}
\details{
The function sets various graphical parameters for ggplot2 plots to customize their appearance.
}
\examples{
set_ggplot_theme()

}
