% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_SoftwareCategory.R
\name{set_SoftwareCategory}
\alias{set_SoftwareCategory}
\title{Set Software Category}
\usage{
set_SoftwareCategory(software, keywords, action, json, session.string)
}
\arguments{
\item{software}{(character) Software name}

\item{keywords}{(character) New keywords}

\item{action}{(character) "add" or "remove" software category}

\item{json}{(list) JSON of all OntoSoft Portal software. Create this list with
\code{get_json()}.}

\item{session.string}{(character) Key generated with \code{login()}.}
}
\value{
(json) Updated JSON object and a JSON file written to \code{tempdir()}.
}
\description{
Add or remove software category keywords.
}
\examples{
\dontrun{
set_keyword('arrow', c('quality control', 'import'))
}

}
