% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourceDir.R
\name{sourceDir}
\alias{sourceDir}
\title{Source all R scripts in a folder}
\usage{
sourceDir(path, trace = TRUE, ...)
}
\arguments{
\item{path}{The directory path to source files from.}
}
\description{
Source all R scripts in a folder
}
