% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_df.R
\name{split_df}
\alias{split_df}
\title{Split Data Frame into Multiple Parts}
\usage{
split_df(df, num_part, return_list = FALSE)
}
\arguments{
\item{df}{A data frame to be split.}

\item{num_part}{An integer specifying the number of columns to split the data frame into.}

\item{return_list}{A logical indicating whether to return a list of data frames (default: FALSE).}
}
\value{
A new data frame that is split into the specified number of columns or a list of data frames if return_list is TRUE.
}
\description{
This function takes a data frame and splits it into multiple parts based on the specified number of columns.
It can also include row names as a column if a valid row name label is provided.
}
\examples{
split_df(mtcars[1:10, 1:3], num_part = 2, return_list = FALSE)
}
