% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepwise_backward_sem_combn.R
\docType{class}
\name{stepwise_backward_sem_combn-class}
\alias{stepwise_backward_sem_combn-class}
\title{Stepwise backward SEM combinatorial}
\arguments{
\item{datos}{a data frame containing the data to be used to fit the model.}

\item{vars_pos}{a character vector containing the names of the variables to consider in the combinations.}
}
\value{
a list with two elements:
\itemize{
\item \code{modelo_final}: a string with the formula of the model selected with the best combination
\item \code{estadisticos_ajuste}: a data frame containing the fit statistics of all combinations evaluated.
}
}
\description{
The function \code{stepwise_backward_sem_combn()} uses the \code{combn()} function to generate all possible combinations of a set of variables (\code{vars_pos}) and then uses \code{sem()} to fit a structural equation model (SEM) with each combination. The goal is to evaluate the fit statistics CFI, TLI and RMSEA for each combination and select the combination with the best fit statistics.
}
\examples{
datos <- your_data_frame
vars_pos <- c("Pos1", "Pos2", "Pos3", "Pos4", "Pos5", "Pos6", "Pos7", "Pos8", "Pos9")
resultados <- stepwise_backward_sem_combn(datos, vars_pos)
modelo_seleccionado <- resultados$modelo_final
estadisticos_ajuste <- resultados$estadisticos_ajuste

}
