% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_plots.R
\name{summarize_plots}
\alias{summarize_plots}
\title{Summarize and Plot Columns of a Data Frame}
\usage{
summarize_plots(df, ncol = 3, fill = "firebrick", width = 10, height = 4, ...)
}
\arguments{
\item{df}{A data frame to summarize and plot.}

\item{ncol}{An integer specifying the number of columns to arrange the plots in (default: 3).
This parameter is useful when displaying plots in a grid.}

\item{fill}{A string specifying the fill color for the bars in histograms or bar plots (default: "firebrick").}

\item{width}{Numeric value specifying the width of the plot (default: 10).}

\item{height}{Numeric value specifying the height of the plot (default: 4).}

\item{...}{Additional arguments passed to \code{patchwork::wrap_plots()} for customizing the layout.}
}
\value{
A combined patchwork of \code{ggplot2} objects arranged in a grid, where each plot corresponds
to one of the columns of the data frame.
}
\description{
This function generates summary plots for each column in a data frame.
For numeric columns, it produces histograms, and for factor or character columns,
it creates bar plots of the frequencies.
}
\details{
This function checks the data type of each column in the provided data frame.
\itemize{
\item For numeric columns, it creates histograms showing the distribution of the values.
\item For categorical columns (factors or characters), it converts any character columns to factors
and generates bar plots showing the counts of each level.
}

The function returns a combined set of plots arranged in a grid using \code{patchwork::wrap_plots}.
You can control the number of columns in the grid using the \code{ncol} argument.
}
\examples{
# Example with the mtcars dataset
summarize_plots(mtcars)

}
