% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_table.R
\name{summarize_table}
\alias{summarize_table}
\title{Summarize Data Frame by Column Type}
\usage{
summarize_table(df)
}
\arguments{
\item{df}{A data frame containing the variables to be summarized.}
}
\value{
A tibble with summary statistics for each variable in the input data frame. The summary includes:
\item{Type}{The data type of the variable (Numeric, Factor, Character, Logical, Other).}
\item{Mean}{The mean of the numeric variables (if applicable).}
\item{Median}{The median of the numeric variables (if applicable).}
\item{SD}{The standard deviation of the numeric variables (if applicable).}
\item{Min}{The minimum value of the numeric variables (if applicable).}
\item{Max}{The maximum value of the numeric variables (if applicable).}
\item{Levels}{The number of levels for factor variables (if applicable).}
\item{Unique}{The number of unique values for character or factor variables.}
\item{Most_Common}{The most frequent value for character or factor variables.}
\item{Frequency}{The frequency of the most common value for character or factor variables.}
\item{True_Count}{The number of TRUE values for logical variables.}
\item{False_Count}{The number of FALSE values for logical variables.}
\item{NA_Count}{The number of missing values for each variable.}
}
\description{
This function provides a summary of each column in a data frame, generating different statistics based on the type of data (numeric, factor, character, logical, or other).
It returns a tibble with the variable names and associated summary statistics, categorized by their data type.
}
\examples{
summarize_table(mtcars)

}
