% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_software_category.R
\name{sync_software_category}
\alias{sync_software_category}
\title{Synchronize software category keywords with the IMCR vocabulary}
\usage{
sync_software_category(name)
}
\arguments{
\item{name}{(character) Software name}
}
\value{
(list) Updated software JSON for the specified \code{name} and added to 
  the \code{imcr_json} object in the global environment.
  (logical) Updated \code{imcr_json_mod_index} object in the global 
  environment, which indicates the specified \code{name} has been modified
  and is used by PUT and POST functions.
}
\description{
Synchronize software category keywords with the IMCR vocabulary
}
\examples{
\dontrun{
get_imcr_json()
modify_software_category("add", "arrow", "thesauri")
sync_software_category("arrow")
}

}
