% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t2m.R
\name{t2m}
\alias{t2m}
\title{Table to matrix}
\usage{
t2m(x)
}
\arguments{
\item{x}{Result from \code{table}}
}
\value{
A matrix.
}
\description{
\code{t2m} Convert \code{table} results to matrix
}
\details{
Converts \code{table} into matrix.
}
\examples{
x1 <- sample(c('a','b'),20,replace = TRUE)
x2 <- sample(c('0','1'),20,replace = TRUE)
t2m(table(x1,x2))
}
