\name{toolkit-package}
\alias{toolkit-package}
\alias{toolkit}
\docType{package}
\title{
Toolkit for data manipulation and plotting.
}
\description{
Toolkit for data manipulation and plotting.
}
\details{
\tabular{ll}{
Package: \tab toolkit\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-02-24\cr
License: \tab What license is it under?\cr
}

}
\author{
Sean Cho

Maintainer: Sean Cho <sean.cho@jhmi.edu>

}

\keyword{ package }
