% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsupervised.R
\name{unsupervised}
\alias{unsupervised}
\title{Unsupervised heatmap by standard deviation}
\usage{
unsupervised(x, n, col = colorRampPalette(c("blue", "black", "yellow"))(50),
  get_hm = FALSE, main = NULL, ...)
}
\arguments{
\item{x}{Numeric matrix with samples in columns and features in rows.}

\item{n}{Number of features.}

\item{col}{Color for heatmap. Defaults to blue-black-yellow color scheme.}

\item{get_hm}{Logical vector whether to return heatmap with feature names.}

\item{main}{Title.}

\item{...}{Passes arguments to \code{heatmap.2}}
}
\value{
A heatmap.2 object and plot.
}
\description{
\code{unsupervised} Performs unsupervised clustering on the most
   variable features. Also returns the names of the features.
}
\details{
Finds the most variable features across all samples by standard
   deviation, clusters them, and plots it on a heatmap using the
   \code{heatmap.2} function.
}
\seealso{
\code{\link[gplots]{heatmap.2}}
}
