% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_software_category.R
\name{update_software_category}
\alias{update_software_category}
\title{Add/remove software category keywords}
\usage{
update_software_category(action, software, keywords, session.string)
}
\arguments{
\item{action}{(character) Action to perform ("add" or "remove")}

\item{software}{(character) Software name}

\item{keywords}{(character) Keywords from the 
\href{http://vocab.lternet.edu/vocab/registry/index.php}{IMCR Controlled Vocabulary}.}

\item{session.string}{(character) Key generated with \code{login()}.}
}
\value{
(list) Updated software JSON represented as a list.
  (.json file) Updated software JSON written to 
  \code{paste0(tempdir(), "/", software, ".json")}.
}
\description{
Add/remove only the narrowest terms, broader terms are automatically 
added/removed.
}
\examples{
\dontrun{
# Get all software JSON from IMCR Portal
json <- get_json("http://imcr.ontosoft.org/repository/software")

# Get session string
sstr <- login()

# Add software category keywords
update_software_category("add", 'arrow', c('quality control', 'import'), sstr)

# Remove software category keywords
update_software_category("remove", 'arrow', c('quality control', 'import'), sstr)
}

}
