% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_SoftwareCategory.R
\name{update_SoftwareCategory}
\alias{update_SoftwareCategory}
\title{Update software category}
\usage{
update_SoftwareCategory(software, json)
}
\arguments{
\item{software}{(character) Software name}

\item{json}{(list) JSON of all OntoSoft Portal software. Create this list with
\code{get_json()}.}

\item{keywords}{(character) New keywords}

\item{action}{(character) "add" or "remove" software category}
}
\value{
(json) Updated JSON object and a JSON file written to \code{tempdir()}.
}
\description{
Update all software category keywords to the current structure of the IMCR
Controlled Vocabulary.
}
\examples{
\dontrun{
update_software_category()
}

}
