% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zpvalue.R
\name{zpvalue}
\alias{zpvalue}
\title{Format p-value for Display}
\usage{
zpvalue(x)
}
\arguments{
\item{x}{A numeric value representing a p-value.}
}
\value{
A character string. If \code{x > 0.001}, it returns the p-value rounded to three decimal places. If \code{x <= 0.001}, it returns '<0.001'.
}
\description{
This function formats a p-value for display, rounding it to three decimal places. If the p-value is smaller than 0.001, it returns '<0.001'.
}
\examples{
# Example usage of the function:
zpvalue(0.005)
zpvalue(0.0005)

}
