% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpFun.R
\name{zround}
\alias{zround}
\title{Round and Format Numbers to Specific Decimal Places}
\usage{
zround(x, digits = 2, method = 1)
}
\arguments{
\item{x}{A numeric vector or a single numeric value.}

\item{digits}{An integer indicating the number of decimal places
to format the numbers. Defaults to 2.}
}
\value{
A character vector with each number in \code{x} formatted to
\code{digits} decimal places.
}
\description{
This function takes a numeric vector or a single numeric value and
formats it to a specified number of decimal places. The function
returns a character vector with the numbers formatted as strings.
}
\examples{

x=c(5.555,1.115,-0.002)
zround(x,method=1)
zround(x,method=2)
formatC(x, digits = 2, format = "f")
formattable::formattable(x, digits = 2, format = "f")

}
