% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip.R
\name{clip}
\alias{clip}
\title{Capture output and copy to clipboard}
\usage{
clip(obj, file = "output.txt")
}
\arguments{
\item{obj}{The R object whose output is to be captured.}

\item{file}{The file to which the output will be written (default: "output.txt").}
}
\value{
None
}
\description{
This function captures the output of an R object and writes it to a file, then copies the contents of the file to the clipboard.
}
