% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_numerical.R
\name{compare_numerical}
\alias{compare_numerical}
\title{Compute Mean, SD for Two Samples Decided by Grouping Variable and Perform One-Way Test}
\usage{
compare_numerical(
  dat,
  var1,
  var2,
  total = FALSE,
  varname,
  colname = "Variables"
)
}
\arguments{
\item{dat}{A data frame containing the variables.}

\item{var1}{The grouping variable (factor or categorical).}

\item{var2}{The numerical variable for which mean and standard deviation are computed.}

\item{varname}{A character string used to name the row for the output table.}
}
\value{
A data frame containing group-wise means and standard deviations of \code{var2} for each level of \code{var1}, along with the p-value from a one-way ANOVA test.
}
\description{
This function calculates the mean and standard deviation of a numerical variable for each level of a grouping variable and performs a one-way ANOVA test. The result is a formatted table that includes group-wise means, standard deviations, and a p-value.
}
\details{
The function computes the mean and standard deviation for \code{var2} within each group defined by \code{var1} and overall. It performs a one-way ANOVA test to assess differences in means between groups. The results are returned in a table that includes the p-value.
}
\examples{
compare_numerical(mtcars,'cyl','mpg',varname = 'mile per gallon')

}
