% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_numerical.R
\name{compare_numerical}
\alias{compare_numerical}
\title{Compute Mean and SD for Two Samples by Grouping Variable and Perform One-Way ANOVA}
\usage{
compare_numerical(
  dat,
  col_var,
  row_var,
  varname = "",
  total = FALSE,
  colname = "Variables"
)
}
\arguments{
\item{dat}{A data frame containing the variables.}

\item{col_var}{The grouping variable (factor or categorical).}

\item{row_var}{The numerical variable for which mean and standard deviation are computed.}

\item{varname}{A character string to label the row in the output table.}

\item{total}{Logical; if TRUE, includes overall mean and SD.}

\item{colname}{A character string for naming the output table column.}
}
\value{
A data frame containing group-wise means and standard deviations of \code{row_var} for each level of \code{col_var}, along with the p-value from the one-way ANOVA test.
}
\description{
This function calculates the mean and standard deviation of a numerical variable for each level of a grouping variable and performs a one-way ANOVA test. The output is a formatted table that includes group-wise means, standard deviations, and a p-value.
}
\details{
The function computes the mean and standard deviation for \code{row_var} within each group defined by \code{col_var}, as well as overall. It performs a one-way ANOVA test to assess differences in means between groups. Results are presented in a table that includes the p-value.
}
\examples{
compare_numerical(mtcars, 'cyl', 'mpg', varname = 'Miles per Gallon')

}
