% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_rnws.R
\name{compile_rnws}
\alias{compile_rnws}
\title{Knit Files to PDF and Track Status}
\usage{
compile_rnws(file_list = NULL)
}
\arguments{
\item{file_list}{A character vector containing the file paths to the Rnw or LaTeX source files. If NULL, the function will search for all \code{.Rnw} files in the current directory.}
}
\value{
A data frame with two columns:
\describe{
\item{file}{The name of the file that was processed.}
\item{success}{Logical value indicating whether the knitting process was successful (\code{TRUE}) or encountered an error (\code{FALSE}).}
}
}
\description{
This function takes a list of Rnw or other LaTeX source files, attempts to knit each one to a PDF using the \code{knit2pdf} function, and tracks the success or failure of each attempt.
}
\details{
The function iterates through the list of files, attempts to compile each one using \code{knit2pdf}, and stores the result (success or failure) in a data frame. If an error occurs during the knitting process, the function will skip that file and proceed to the next one.
}
\examples{
\dontrun{
  file_list = list.files(pattern = "^Sta.*Rnw", ignore.case = TRUE)
  knit_results <- compile_rnws(file_list)
  print(knit_results)
}

}
