% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_software.R
\name{delete_software}
\alias{delete_software}
\title{Delete software from portal}
\usage{
delete_software(software.name, session.string, json)
}
\arguments{
\item{software.name}{(character) Software name}

\item{session.string}{(character) Key generated with \code{login()}.}

\item{json}{(list) JSON of all portal software. Create this list with
\code{get_json()}.}
}
\value{
(message) Status message of operation.
}
\description{
Delete software from portal
}
\examples{
\dontrun{
sstr <- login()
delete_software('arrow', sstr)
}

}
