% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_software.R
\name{delete_software}
\alias{delete_software}
\title{Delete software from portal}
\usage{
delete_software(software.name)
}
\arguments{
\item{software.name}{(character) Software name}
}
\value{
(message) Status message of operation.
}
\description{
Delete software from portal
}
\examples{
\dontrun{
login()
delete_software('arrow')
}

}
