% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_broader_terms.R
\name{get_broader_terms}
\alias{get_broader_terms}
\title{Get broader terms for a given IMCR Vocabulary keyword}
\usage{
get_broader_terms(term)
}
\arguments{
\item{keyword}{(character) IMCR Vocabulary keyword}
}
\value{
(character) The input term and any broader terms related to it. A warning
  is issued if the term was not found.
}
\description{
Get broader terms for a given IMCR Vocabulary keyword
}
\examples{
\dontrun{
get_broader_terms("workflow planning")
}

}
