% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_imcr_json.R
\name{get_imcr_json}
\alias{get_imcr_json}
\title{Get JSON of all software in the IMCR Portal}
\usage{
get_imcr_json()
}
\value{
(list) List of JSON metadata for all software in the IMCR Portal. This 
  object is named "imcr_json" and is returned to the global environment 
  where it is accessed by other \code{toolkit} functions.
}
\description{
Get JSON of all software in the IMCR Portal
}
\examples{
\dontrun{
get_imcr_json()
}

}
