% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_json.R
\name{get_json}
\alias{get_json}
\title{Get JSON metadata for all portal software}
\usage{
get_json(url = "http://imcr.ontosoft.org/repository/software")
}
\arguments{
\item{url}{(character) OntoSoft portal URL.}
}
\value{
(list) Named list of JSON metadata for all software in a specified 
  OntoSoft Portal. This object is named "json" and is returned to the global
  environment for use with other toolkit functions.
}
\description{
Get JSON metadata for all portal software
}
\examples{
\dontrun{
get_json("http://imcr.ontosoft.org/repository/software")
}

}
