% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_json.R
\name{get_json}
\alias{get_json}
\title{Get JSON metadata for all portal software}
\usage{
get_json(url)
}
\arguments{
\item{url}{(character) Portal URL}
}
\value{
(list) JSON metadata for all portal software
}
\description{
Get JSON metadata for all portal software
}
\examples{
\dontrun{
json <- get_json('http://imcr.ontosoft.org/repository/software')
}

}
