% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{login}
\alias{login}
\title{Login}
\usage{
login(user.name = NULL, user.pass = NULL)
}
\arguments{
\item{user.name}{(character) OntoSoft user name. Default (NULL) queries for this
information in the console.}

\item{user.pass}{(character) OntoSoft user password. Default (NULL) queries for this
information in the console.}
}
\value{
(character) Session string object named "imcr_session_string" and created
  in the global environment for use with other \code{toolkit} functions.
}
\description{
Login
}
\examples{
\dontrun{
login()
}

}
