% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpFun.R
\name{meanSD}
\alias{meanSD}
\title{Calculate mean and SD of a vector}
\usage{
meanSD(x, digits = 2)
}
\arguments{
\item{digits}{number of digits after the decimal point. Default is 2.}
}
\description{
Calculate mean and SD of a vector
}
\examples{
meanSD(rnorm(100))
}
