% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_software_category.R
\name{modify_software_category}
\alias{modify_software_category}
\title{Add/remove software category keywords}
\usage{
modify_software_category(action, name, term)
}
\arguments{
\item{action}{(character) "add" or "remove" keywords}

\item{name}{(character) Software name}

\item{term}{(character) Keywords from the 
\href{http://vocab.lternet.edu/vocab/registry/index.php}{IMCR Controlled Vocabulary}.}
}
\value{
(list) Updated software JSON for the specified \code{name} and added to 
  the \code{imcr_json} object in the global environment.
  (logical) Updated \code{imcr_json_mod_index} object in the global 
  environment, which indicates the specified \code{name} has been modified
  and is used by PUT and POST functions.
}
\description{
Add/remove software category keywords
}
\examples{
\dontrun{
get_imcr_json()
modify_software_category("add", "arrow", c("quality control", "import"))
modify_software_category("remove", "arrow", "quality control")
}

}
