% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointbox.R
\name{pointbox}
\alias{pointbox}
\title{Boxplots with points}
\usage{
pointbox(value, grp, b_col = NULL, s_col = NULL, s_pch = 16,
  s_jitter = NULL, showN = TRUE, heightN = NULL, stat_test = TRUE,
  y_range = NULL, s_outline = TRUE, ...)
}
\arguments{
\item{value}{Numeric variable for N measurements.}

\item{grp}{Classification of measurements.}

\item{b_col}{Boxplot color. Default: NULL (black)}

\item{s_col}{Scatterplot color. Default: NULL (black). If specified, length of N, 
as each point can be colored separately.}

\item{s_pch}{Point type. Default: 16, filled circle.}

\item{s_jitter}{Jitter factor for points.}

\item{showN}{Logical. Show number of samples in each classification.}

\item{heightN}{Manually set height for N if showN = TRUE.}

\item{stat_test}{Logical. Perform statistical test? Currently, only one-sided
anova is implemented.}

\item{y_range}{Manually specify y-axis range.}

\item{s_outline}{Logical. Outline points.}

\item{...}{Arguments passed on to par. See \code{\link[graphics]{par}}.}
}
\value{
Boxplot with horizontally jittered points.
}
\description{
\code{pointbox} Plot boxplots with points (scatterplot).
\code{pointbox_label} Plot boxplots with points and labels each point.
}
\details{
Colors for boxes and points are defined separately to allow for
   fine-tuned user control.
}
\examples{
# Make data
d <- c(rnorm(n = 10, mean = 1, sd = 0.5),
       rnorm(n = 10, mean = 5, sd = 0.5))
g <- c(rep('a', 10), rep('b', 10))
pointbox(value = d, grp = g)

}
\seealso{
\code{\link[graphics]{par}}
}

