% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_software.R
\name{post_software}
\alias{post_software}
\title{Add new software}
\usage{
post_software(software.name, session.string, path)
}
\arguments{
\item{software.name}{(character) Software name}

\item{session.string}{(character) Key generated with \code{login()}.}

\item{path}{(list) Path to software JSON.}
}
\value{
(message) Status message of operation.
}
\description{
Add new software
}
\examples{
\dontrun{
sstr <- login()
post_software('arrow', sstr, '/path/to/software')
}

}
