% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_software.R
\name{post_software}
\alias{post_software}
\title{Add new software to the IMCR Portal}
\usage{
post_software(name, path)
}
\arguments{
\item{name}{(character) File name of software JSON (without file extension).}

\item{path}{(character) Path to software JSON.}
}
\value{
(message) Status message of POST operation.
}
\description{
Add new software to the IMCR Portal
}
\examples{
\dontrun{
login()
post_software("arrow", "/path/to/software/json")
}

}
