% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_flextable.R
\name{print_flextable}
\alias{print_flextable}
\title{Print a Styled Table for HTML or Word Output}
\usage{
print_flextable(
  df,
  num_col = 1,
  rowname_label = NA,
  caption = "",
  digits = 2,
  fontsize = 11,
  big.mark = ",",
  na_str = "",
  ...
)
}
\arguments{
\item{df}{The data frame to be formatted as a table.}

\item{num_col}{An integer specifying the number of columns to split the output table into (default: 1).}

\item{rowname_label}{A character string specifying the column name for the row names; if not provided, the row names will not be printed (default: NA).}

\item{caption}{A character string specifying the caption for the table (default: '').}

\item{digits}{An integer indicating the number of digits to display (default: 2).}

\item{fontsize}{Font size for the table (default: 9).}

\item{big.mark}{A character string used as the thousands separator for numbers (default: ',').}

\item{na_str}{A string used to replace missing values (default: '').}

\item{...}{Additional arguments to be passed to the formatting functions.}
}
\value{
A styled table in HTML or Word format.
}
\description{
This function prints a styled table in HTML or Word format using flextable.
}
\examples{
dt = mtcars[1:10, 1:3]
print_flextable(dt, num_col = 2, rowname_label = 'car')

}
