% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_table.R
\name{print_table}
\alias{print_table}
\title{Print a styled table for HTML or LaTeX output}
\usage{
print_table(
  df,
  caption = "",
  digits = 2,
  print_rownames = FALSE,
  fontsize = 9,
  ...
)
}
\arguments{
\item{df}{The data frame to be formatted as a table.}

\item{caption}{The caption for the table (default: '').}

\item{digits}{Number of digits to display (default: 2).}

\item{print_rownames}{Logical indicating whether to print row names (default: FALSE).}

\item{fontsize}{Font size for the table (default: 9).}

\item{...}{Additional arguments to be passed to the formatting functions.}
}
\value{
A styled table either in HTML or LaTeX format.
}
\description{
This function prints a styled table either in HTML or LaTeX format based on the output document type.
}
\examples{

}
