% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put_software.R
\name{put_software}
\alias{put_software}
\title{Update software in the IMCR Portal}
\usage{
put_software()
}
\value{
(message) Status message of PUT operation.
}
\description{
Modified software metadata (indicated by a "TRUE" in the 
\code{imcr_json_mod_index} object) is written to \code{tempdir()} and 
uploaded to the IMCR Portal.
}
\examples{
\dontrun{
login()
get_imcr_json()
modify_software_category("add", "arrow", c("quality control", "import"))
put_software()
}

}
