% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put_software.R
\name{put_software}
\alias{put_software}
\title{Update software}
\usage{
put_software(software.name, path)
}
\arguments{
\item{software.name}{(character) Software name}

\item{path}{(list) Path to software JSON.}
}
\value{
(message) Status message of operation.
}
\description{
Update software
}
\examples{
\dontrun{
sstr <- login()
put_software('arrow', sstr, '/path/to/software')
}

}
