% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_software_category.R
\name{set_software_category}
\alias{set_software_category}
\title{Add/remove software category keywords}
\usage{
set_software_category(action, software, keywords, session.string)
}
\arguments{
\item{action}{(character) "add" or "remove" software category keywords}

\item{software}{(character) Software name}

\item{keywords}{(character) Software category keywords}

\item{session.string}{(character) Key generated with \code{login()}.}
}
\value{
(json file) Software JSON file written to 
  \code{paste0(tempdir(), '/', software)}.
}
\description{
Only the narrowest terms need to 
be added, \code{align_software_category()} automatically adds and removes
broader terms of a specified vocabulary.
}
\examples{
\dontrun{
# Get all software JSON from IMCR Portal
json <- get_json("http://imcr.ontosoft.org/repository/software")

# Get session string
sstr <- login()

# Add software category keywords
set_software_category("add", 'arrow', c('quality control', 'import'), sstr)

# Remove software category keywords
set_software_category("remove", 'arrow', c('quality control', 'import'), sstr)
}

}
