% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepwise_backward_sem_combn.R
\docType{class}
\name{stepwise_backward_sem_combn-class}
\alias{stepwise_backward_sem_combn-class}
\title{Stepwise backward selection of SEM combinations}
\arguments{
\item{datos}{A data frame containing the data}

\item{vars_pos}{A character vector containing the variables to consider in the analysis.}
}
\value{
An object of class stepwise_backward_sem_combn which contains the final model, fit statistics and input data.
}
\description{
A class for performing stepwise backward selection of variables in SEM models and storing the final model and fit statistics.
}
\note{
The function uses the 'lavaan', 'ggplot2', and 'cowplot' package. Make sure to have it installed before using the function.
}
\examples{
data(holzingerSwineford1939)
vars_pos <- names(holzingerSwineford1939)[1:8]
modelo <- stepwise_backward_sem_combn(holzingerSwineford1939, vars_pos)
print(modelo)
plot(modelo)
}
