% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_software_category.R
\name{sync_software_category}
\alias{sync_software_category}
\title{Synchronize software category keywords with the IMCR vocabulary}
\usage{
sync_software_category(software)
}
\value{
(list) Updated software JSON as a list object.
  (.json file) Updated software JSON written to the IMCR Portal, and to
  \code{paste0(tempdir(), "/", software, ".json")}.
}
\description{
Synchronize software category keywords with the IMCR vocabulary
}
\examples{
\dontrun{
login()
sync_software_category()
}

}
