% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolkit-package.R
\docType{package}
\name{toolkit-package}
\alias{toolkit}
\alias{toolkit-package}
\title{toolkit: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Guangyu Zhu \email{guangyuzhu@uri.edu}

}
\keyword{internal}
