% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsupervised3.R
\name{unsupervised.3}
\alias{unsupervised.3}
\title{Unsupervised heatmap by standard deviation}
\usage{
unsupervised.3(x, n, col = colorRampPalette(c("blue", "black", "yellow"))(50),
  get_hm = FALSE, main = NULL, ...)
}
\arguments{
\item{x}{Numeric matrix with samples in columns and features in rows.}

\item{n}{Number of features.}

\item{col}{Color for heatmap. Defaults to blue-black-yellow color scheme.}

\item{get_hm}{Logical vector whether to return heatmap with feature names.}

\item{main}{Title.}

\item{...}{Passes arguments to \code{heatmap.3}}
}
\value{
A heatmap.3 object and plot.
}
\description{
\code{unsupervised.3} Performs unsupervised clustering on the most
   variable features using heatmap.3 which allows for multiple Row/ColSideCol.
}
\details{
Finds the most variable features across all samples by standard
   deviation, clusters them, and plots it on a heatmap using the
   \code{heatmap.3} function.
}
