% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_software_category.R
\name{update_software_category}
\alias{update_software_category}
\title{Add/remove software category keywords}
\usage{
update_software_category(action, software, keywords)
}
\arguments{
\item{action}{(character) Action to perform ("add" or "remove")}

\item{software}{(character) Software name}

\item{keywords}{(character) Keywords from the 
\href{http://vocab.lternet.edu/vocab/registry/index.php}{IMCR Controlled Vocabulary}.}
}
\value{
(list) Updated software JSON as a list object.
  (.json file) Updated software JSON written to the IMCR Portal, and to
  \code{paste0(tempdir(), "/", software, ".json")}.
}
\description{
Only the narrowest terms need to be input, broader terms are automatically 
added/removed.
}
\examples{
\dontrun{
# Get all IMCR JSON
get_imcr_json()

# Login
login()

# Add software category keywords
update_software_category("add", 'arrow', c('quality control', 'import'))

# Remove software category keywords
update_software_category("remove", 'arrow', c('quality control', 'import'))
}

}
