% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_software_category.R
\name{update_software_category}
\alias{update_software_category}
\title{Update software category}
\usage{
update_software_category(software, json)
}
\arguments{
\item{software}{(character) Software name}

\item{json}{(list) JSON of all OntoSoft Portal software. Create this list with
\code{get_all_software()}.}

\item{keywords}{(character) New keywords}

\item{method}{(character) "add" (add new keyword), "remove" (remove existing keyword),
or "align" (align with a controlled vocabulary.}
}
\value{
(json) Updated JSON object and a JSON file written to \code{tempdir()}.
}
\description{
Add, remove, and align software category keywords.
}
\examples{
\dontrun{
# Add new keyword and all broader terms
update_software_category(
  'arrow',
  c('workflows', 'loading'),
  'add',
  json
)
}

}
