% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zkbl.R
\name{zkbl}
\alias{zkbl}
\title{Create a styled LaTeX table using kableExtra with additional options}
\usage{
zkbl(table, booktabs = TRUE, escape = TRUE, ...)
}
\arguments{
\item{table}{The data table to be formatted.}

\item{booktabs}{Logical indicating whether to use the booktabs formatting (default: TRUE).}

\item{escape}{Logical indicating whether to escape special characters in the table (default: TRUE).}

\item{...}{Additional arguments to be passed to the kbl function.}
}
\value{
A LaTeX table formatted using kableExtra package.
}
\description{
This function creates a LaTeX table using the kableExtra package with the provided options.
}
\examples{
library(kableExtra)
df <- data.frame(x = 1:5, y = letters[1:5])
zkbl(df)

}
