% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpFun.R
\name{zround}
\alias{zround}
\title{Rounding of Numbers to a string}
\usage{
zround(x, digits = 2)
}
\arguments{
\item{x}{a numeric vector}

\item{digits}{number of digits after the decimal point. Default is 2.}
}
\description{
Takes a single numeric argument x and returns a character vector containing strings rounded to specified number of digits after the decimal point (default 2
}
\examples{
zround(c(pi,pi*10^5),3)
}
